+++ definitions/SpaceDefinitions.php
@@ -11,13 +11,11 @@ use humhub\modules\space\models\Membership;
 use humhub\modules\space\models\Space;
 use yii\helpers\Url;
 
-
 /**
  * Class SpaceDefinitions
  */
 class SpaceDefinitions
 {
-
     public static function getSpaceShort(Space $space)
     {
         return [
@@ -25,7 +23,7 @@ class SpaceDefinitions
             'guid' => $space->guid,
             'name' => $space->name,
             'description' => $space->description,
-            'url' => Url::to(['/', 'container' => $space], true)
+            'url' => Url::to(['/', 'container' => $space], true),
         ];
     }
 
@@ -70,4 +68,3 @@ class SpaceDefinitions
     }
 
 }
-