+++ messages/es-419/base.php
@@ -1,0 +1,22 @@
+<?php
+
+return [
+    '<strong>REST</strong> Configuration' => '',
+    'Access Tokens' => '',
+    'Add Access Token' => '',
+    'Allow Bearer Authentication' => '',
+    'Allow HTTP Basic Authentication' => '',
+    'Allow JWT Authentication' => '',
+    'Allow Query Param Bearer Authentication' => '',
+    'Bearer Access Token Successfully Revoked' => '',
+    'Bearer Auth' => '',
+    'Enabled for all registered users' => '',
+    'Expiration' => '',
+    'General' => '',
+    'JWT Auth' => '',
+    'JWT Key' => '',
+    'JWT Token Expiration' => '',
+    'No enabled modules found with additional REST API endpoints.' => '',
+    'Token' => '',
+    'User' => '',
+];