+++ migrations/uninstall.php
@@ -4,7 +4,6 @@ use yii\db\Migration;
 
 class uninstall extends Migration
 {
-
     public function up()
     {
         $this->dropTable('impersonate_auth_tokens');